Check if a directory exist before executing `Directory.rmdir`. Since Node.js v16 it generate an error if the directory doesn't exist

+++ src/Nacara.Core/Node.Extra.fs
@@ -104,13 +104,18 @@ module Directory =
             "recursive" ==> true
         ]
 
-        Promise.create (fun resolve reject ->
-            fs?rmdir(dir, options, (fun (err : Base.ErrnoException option) ->
-                match err with
-                | Some err -> reject (err :?> System.Exception)
-                | None -> resolve ()
-            ))
-        )
+        promise {
+            let! dirExist = exists dir
+
+            if dirExist then
+                do! Promise.create (fun resolve reject ->
+                    fs?rmdir(dir, options, (fun (err : Base.ErrnoException option) ->
+                        match err with
+                        | Some err -> reject (err :?> System.Exception)
+                        | None -> resolve ()
+                    ))
+                )
+        }
 
 [<RequireQualifiedAccess>]
 module File =
